@@ -1,132 +1,134 @@
 {
-  "config": {
-    "optimize-autoloader": true,
-    "platform": {
-      "php": "7.4"
+    "config": {
+        "optimize-autoloader": true,
+        "platform": {
+            "php": "7.4"
+        },
+        "preferred-install": "dist",
+        "sort-packages": true
     },
-    "preferred-install": "dist",
-    "sort-packages": true
-  },
-  "name": "laravel/laravel",
-  "type": "project",
-  "description": "The Laravel Framework.",
-  "keywords": [
-    "framework",
-    "laravel"
-  ],
-  "license": "MIT",
-  "repositories": [
-    {
-      "type": "vcs",
-      "url": "https://github.com/pcartagena74/ticketit.git"
-    }
-  ],
-  "require": {
-    "alexusmai/laravel-file-manager": "^2.4",
-    "asdfx/laravel-fullcalendar": "^1.2",
-    "barryvdh/laravel-snappy": "^0.4.4",
-    "bugsnag/bugsnag-laravel": "^2.22",
-    "doctrine/dbal": "^3.1",
-    "eduardokum/laravel-mail-auto-embed": "^1.0",
-    "fedeisas/laravel-mail-css-inliner": "^3.0",
-    "fennb/phirehose": "^1.0",
-    "fideloper/proxy": "^4.4",
-    "google/recaptcha": "^1.2",
-    "graham-campbell/flysystem": "^7.1",
-    "h4cc/wkhtmltoimage-amd64": "0.12.x",
-    "h4cc/wkhtmltopdf-amd64": "dev-master#4144bccbfaa40e8150a763e2d55dfedce9d13909",
-    "intervention/image": "^2.6",
-    "jdavidbakr/mail-tracker": "^2.1",
-    "jenssegers/agent": "^2.6",
-    "knplabs/knp-snappy": "^1.1",
-    "kordy/ticketit": "dev-master",
-    "kriswallsmith/buzz": "^1.1",
-    "laravel/cashier": "^12.15",
-    "laravel/dusk": "^6.17",
-    "laravel/framework": "^7.29",
-    "laravel/socialite": "^5.2",
-    "laravel/telescope": "^3.5",
-    "laravel/tinker": "^2.5",
-    "laravelcollective/html": "^6.2",
-    "league/flysystem": "^1.0",
-    "league/flysystem-aws-s3-v3": "^1.0",
-    "maatwebsite/excel": "^3.1",
-    "mailgun/mailgun-php": "^3.0",
-    "moontoast/math": "1.1.2 as 1.999.999",
-    "nyholm/psr7": "^1.2",
-    "paquettg/php-html-parser": "^2.2",
-    "phpmd/phpmd": "^2.6",
-    "predis/predis": "^1.1",
-    "rap2hpoutre/fast-excel": "^3.0",
-    "socialiteproviders/linkedin": "^3.0",
-    "spatie/browsershot": "^3.49",
-    "spatie/laravel-activitylog": "^3.17",
-    "spatie/laravel-referer": "1.7.1",
-    "squizlabs/php_codesniffer": "^3.6",
-    "stripe/stripe-php": "^7.94",
-    "shanmuga/laravel-entrust": "^2.2",
-    "laravel/ui": "^2.5",
-    "fruitcake/laravel-cors": "^2.0",
-    "guzzlehttp/guzzle": "^6.3.1|^7.0.1",
-    "php": "^7.2.5|^8.0"
-  },
-  "require-dev": {
-    "barryvdh/laravel-debugbar": "^3.6",
-    "mockery/mockery": "^1.3.1",
-    "nunomaduro/collision": "^4.3",
-    "phpunit/phpunit": "^8.5.8|^9.3.3",
-    "fakerphp/faker": "^1.9.1",
-    "facade/ignition": "^2.0"
-  },
-  "extra": {
-    "laravel": {
-      "dont-discover": [
-        "barryvdh/laravel-debugbar"
-      ]
-    }
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": [
-        "app/"
-      ]
+    "name": "laravel/laravel",
+    "type": "project",
+    "description": "The Laravel Framework.",
+    "keywords": [
+        "framework",
+        "laravel"
+    ],
+    "license": "MIT",
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "https://github.com/pcartagena74/ticketit.git"
+        }
+    ],
+    "require": {
+        "alexusmai/laravel-file-manager": "^2.4",
+        "asdfx/laravel-fullcalendar": "^1.2",
+        "barryvdh/laravel-snappy": "^0.4.4",
+        "bugsnag/bugsnag-laravel": "^2.22",
+        "doctrine/dbal": "^3.1",
+        "eduardokum/laravel-mail-auto-embed": "^1.0",
+        "fedeisas/laravel-mail-css-inliner": "^3.0",
+        "fennb/phirehose": "^1.0",
+        "google/recaptcha": "^1.2",
+        "graham-campbell/flysystem": "^7.1",
+        "h4cc/wkhtmltoimage-amd64": "0.12.x",
+        "h4cc/wkhtmltopdf-amd64": "dev-master#4144bccbfaa40e8150a763e2d55dfedce9d13909",
+        "intervention/image": "^2.6",
+        "jdavidbakr/mail-tracker": "^2.1",
+        "jenssegers/agent": "^2.6",
+        "knplabs/knp-snappy": "^1.1",
+        "kordy/ticketit": "dev-master",
+        "kriswallsmith/buzz": "^1.1",
+        "laravel/cashier": "^13.4",
+        "laravel/dusk": "^6.17",
+        "laravel/framework": "^8.55",
+        "laravel/socialite": "^5.2",
+        "laravel/telescope": "^4.6",
+        "laravel/tinker": "^2.5",
+        "laravelcollective/html": "^6.2",
+        "league/flysystem": "^1.0",
+        "league/flysystem-aws-s3-v3": "^1.0",
+        "maatwebsite/excel": "^3.1",
+        "mailgun/mailgun-php": "^3.0",
+        "moontoast/math": "1.1.2 as 1.999.999",
+        "nyholm/psr7": "^1.2",
+        "paquettg/php-html-parser": "^2.2",
+        "phpmd/phpmd": "^2.6",
+        "predis/predis": "^1.1",
+        "rap2hpoutre/fast-excel": "^3.0",
+        "socialiteproviders/linkedin": "^3.0",
+        "spatie/browsershot": "^3.49",
+        "spatie/laravel-activitylog": "^3.17",
+        "spatie/laravel-referer": "1.7.1",
+        "squizlabs/php_codesniffer": "^3.6",
+        "stripe/stripe-php": "^7.94",
+        "shanmuga/laravel-entrust": "^2.2",
+        "laravel/ui": "^3.3",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^6.3.1|^7.0.1",
+        "php": "^7.3|^8.0"
     },
-    "exclude-from-classmap": [
-      "vendor/kordy/ticketit/src/ViewComposers/TicketItComposer.php",
-      "vendor/kordy/ticketit/src/TicketitServiceProvider.php",
-      "vendor/alexusmai/laravel-file-manager/src/FileManager.php"
-    ],
-    "classmap": [
-      "database/seeds",
-      "database/factories",
-      "app/Models"
-    ],
-    "files": [
-      "app/Helpers/helper_functions.php",
-      "app/Vendor/Ticketit/TicketItComposer.php",
-      "app/Vendor/Ticketit/TicketitServiceProvider.php",
-      "app/Vendor/Maatwebsite/Excel/ReadChunk.php",
-      "app/Vendor/Alexusmai/LaravelFileManager/FileManager.php"
-    ]
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Tests\\": "tests/",
-      "vendor\\kordy\\ticketit\\src\\ViewComposers\\": "app/Vendor/Ticketit/"
+    "require-dev": {
+        "barryvdh/laravel-debugbar": "^3.6",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^8.5.8|^9.3.3",
+        "fakerphp/faker": "^1.9.1",
+        "facade/ignition": "^2.5"
+    },
+    "extra": {
+        "laravel": {
+            "dont-discover": [
+                "barryvdh/laravel-debugbar"
+            ]
+        }
+    },
+    "autoload": {
+        "psr-4": {
+            "App\\": [
+                "app/"
+            ],
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        },
+        "exclude-from-classmap": [
+            "vendor/kordy/ticketit/src/ViewComposers/TicketItComposer.php",
+            "vendor/kordy/ticketit/src/TicketitServiceProvider.php",
+            "vendor/alexusmai/laravel-file-manager/src/FileManager.php"
+        ],
+        "classmap": [
+            "app/Models"
+        ],
+        "files": [
+            "app/Helpers/helper_functions.php",
+            "app/Vendor/Ticketit/TicketItComposer.php",
+            "app/Vendor/Ticketit/TicketitServiceProvider.php",
+            "app/Vendor/Maatwebsite/Excel/ReadChunk.php",
+            "app/Vendor/Alexusmai/LaravelFileManager/FileManager.php"
+        ]
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/",
+            "vendor\\kordy\\ticketit\\src\\ViewComposers\\": "app/Vendor/Ticketit/"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "scripts": {
+        "post-autoload-dump": [
+            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
+            "@php artisan package:discover --ansi"
+        ],
+        "post-root-package-install": [
+            "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
+        ],
+        "post-create-project-cmd": [
+            "@php artisan key:generate --ansi"
+        ],
+        "post-update-cmd": [
+            "@php artisan vendor:publish --tag=laravel-assets --ansi"
+        ]
     }
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true,
-  "scripts": {
-    "post-autoload-dump": [
-      "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-      "@php artisan package:discover --ansi"
-    ],
-    "post-root-package-install": [
-      "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
-    ],
-    "post-create-project-cmd": [
-      "@php artisan key:generate --ansi"
-    ]
-  }
 }
