@@ -16,7 +16,10 @@ class Address extends Model
     protected $primaryKey = 'addrID';
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'deleted_at', 'updateDate'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
 
     public function person()
     {
