@@ -11,7 +11,10 @@ class AdminPropGroup extends Model
     protected $primaryKey = 'id';
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'updateDate'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
 
     public function props()
     {
