@@ -14,7 +14,10 @@ class Bundle extends Model
     protected $primaryKey = '';
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'updateDate'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
 
     protected static $logOnlyDirty = true;
     protected static $submitEmptyLogs = false;
