@@ -17,7 +17,10 @@ class EventDiscount extends Model
     protected $primaryKey = 'discountID';
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'updateDate', 'deleted_at'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
 
     protected static $logAttributes = ['percent', 'flatAmt', 'discountCODE'];
     protected static $ignoreChangedAttributes = ['createDate'];
