@@ -12,7 +12,10 @@ class OrgAdminProp extends Model
     //protected $primaryKey = ['orgID', 'propID'];
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'updateDate'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
     protected $fillable = ['orgID', 'propID', 'value'];
 
     protected function setKeysForSaveQuery(Builder $query)
