@@ -14,7 +14,10 @@ class PersonSocialite extends Model
     protected $primaryKey = 'id';
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'updateDate', 'deleted_at'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
 
     public function person()
     {
