@@ -14,7 +14,11 @@ class PersonStaging extends Model
     protected $primaryKey = 'personID';
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
-    protected $dates = ['createDate', 'deleted_at', 'updateDate', 'lastLoginDate'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+        'lastLoginDate' => 'datetime',
+    ];
 
     //protected static $logAttributes = ['prefName', 'login', 'defaultOrgID', 'title', 'compName', 'indName', 'allergenInfo', 'affiliation'];
     //protected static $ignoreChangedAttributes = ['createDate'];
