@@ -13,7 +13,10 @@ class Track extends Model
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
     protected $primaryKey = 'trackID';
-    protected $dates = ['createDate', 'updateDate', 'deleted_at'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+    ];
 
     public function event()
     {
