@@ -16,7 +16,13 @@ class User extends Authenticatable implements \Illuminate\Contracts\Auth\CanRese
     use EntrustUserTraitOver;
 
     protected $table = 'users';
-    protected $dates = ['createDate', 'updateDate', 'trial_ends_at', 'subscription_ends_at', 'last_login'];
+    protected $casts = [
+        'createDate' => 'datetime',
+        'updateDate' => 'datetime',
+        'trial_ends_at' => 'datetime',
+        'subscription_ends_at' => 'datetime',
+        'last_login' => 'datetime',
+    ];
 
     const CREATED_AT = 'createDate';
     const UPDATED_AT = 'updateDate';
