@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use Entrust;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Route;
+use Closure;
 
 class HomeController extends Controller
 {
@@ -13,7 +16,7 @@ class HomeController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('guest');
+        $this->middleware('guest')->except('showApplicationRoutes');
     }
 
     /**
@@ -31,5 +34,55 @@ class HomeController extends Controller
         // This is the function that processes issues reported by error page
         // Responds to POST /reportissue
         dd($request);
+    }
+
+    /**
+     * Show application routes.
+     *
+     * Forbidden in production environment.
+     *
+     * @return \Illuminate\View\View
+     */
+    public function showApplicationRoutes()
+    {
+        /*
+        if (config('app.log_level') == 'production') {
+            abort(403);
+        }
+        */
+
+        if (Entrust::hasRole('Developer')) {
+            $routes = collect(Route::getRoutes());
+
+            $routes = $routes->map(function ($route) {
+                return [
+                    'host' => $route->action['where'],
+                    'uri' => $route->uri,
+                    'name' => $route->action['as'] ?? '',
+                    'methods' => $route->methods,
+                    'action' => $route->action['controller'] ?? 'Closure',
+                    'middleware' => $this->getRouteMiddleware($route),
+                ];
+            });
+
+            return view('v1.auth_pages.admin.routes', compact('routes'));
+        } else {
+            request()->session()->flash('alert-warning', trans('ticketit::lang.you-are-not-permitted-to-access'));
+            return back();
+        }
+    }
+
+    /**
+     * Get route middleware.
+     *
+     * @param \Illuminate\Routing\Route $route
+     *
+     * @return string
+     */
+    protected function getRouteMiddleware($route)
+    {
+        return collect($route->gatherMiddleware())->map(function ($middleware) {
+            return $middleware instanceof Closure ? 'Closure' : $middleware;
+        })->implode(', ');
     }
 }
