@@ -31,7 +31,7 @@ class AdministratorsController extends Controller
 
         Session::flash('status', trans('ticketit::lang.administrators-are-added-to-administrators', ['names' => $administrators_names]));
 
-        return redirect()->action('\Kordy\Ticketit\Controllers\AdministratorsController@index');
+        return redirect()->action([self::class, 'index']);
     }
 
     public function update($id, Request $request)
@@ -40,7 +40,7 @@ class AdministratorsController extends Controller
 
         Session::flash('status', trans('ticketit::lang.administrators-joined-categories-ok'));
 
-        return redirect()->action('\Kordy\Ticketit\Controllers\AdministratorsController@index');
+        return redirect()->action([self::class, 'index']);
     }
 
     public function destroy($id)
@@ -49,7 +49,7 @@ class AdministratorsController extends Controller
 
         Session::flash('status', trans('ticketit::lang.administrators-is-removed-from-team', ['name' => $administrator->name]));
 
-        return redirect()->action('\Kordy\Ticketit\Controllers\AdministratorsController@index');
+        return redirect()->action([self::class, 'index']);
     }
 
     /**
