@@ -285,6 +285,6 @@ class InstallControllerOver extends Controller
         $seeder->run();
         session()->flash('status', 'Demo tickets, users, and agents are seeded!');
 
-        return redirect()->action('\Kordy\Ticketit\Controllers\TicketsController@index');
+        return redirect()->action([\Kordy\Ticketit\Controllers\TicketsController::class, 'index']);
     }
 }
